requirements.txt
beautifulsoup4==4.9.3
feedparser==6.0.2
mysql-connector-python==8.0.23
protobuf==3.15.8
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
